[FIX] base/0.0.0: migrate ir_translation only if it exists
The ir_translation table doesn't exist anymore in 16.0, so the upgrade
from 16 to master will always fail as multiple scripts try to clean up
various bits of ir_translation.

Gate them on whether the table exists.

Some of these scripts may need to be reimplemented on top of the new
translations implementation (rm-unpaired seems like it still makes
sense in the new model).

+++ migrations/base/0.0.0/pre-clean-translations.py
@@ -1,7 +1,11 @@
 # -*- coding: utf-8 -*-
+from odoo.addons.base.maintenance.migrations import util
 
 
 def migrate(cr, version):
+    if not util.table_exists(cr, "ir_translation"):
+        return
+
     cr.execute("DELETE FROM ir_translation WHERE value = src AND type IS DISTINCT FROM 'model'")
     cr.execute("DELETE FROM ir_translation WHERE value IS NULL")
     cr.execute("DELETE FROM ir_translation WHERE value = ''")
